Updated custom form row
